pyproject.toml
[tool.poetry]
name = "manage-employers-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
gunicorn = "^21.0.1"
whitenoise = "^6.5.0"
djangorestframework = "^3.14.0"

[tool.poetry.scripts]
start = "runserver:djtest"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
